+++ src/__fixtures__/responses/permanentDeleteSuccessResponse.fixture.ts
@@ -1,0 +1,18 @@
+export const permanentDeleteSuccessResponseFixture = {
+  latestClientVersion: '',
+  methodResponses: [
+    [
+      'MaskedEmail/set',
+      {
+        accountId: 'u7cb112016d',
+        destroyed: ['masked-81873752'],
+        oldState: null,
+        newState: null,
+        updated: {},
+        created: {}
+      },
+      'a'
+    ]
+  ],
+  sessionState: 'cyrus-68;p-b14967af44'
+};