+++ src/__tests__/get.test.ts
@@ -3,11 +3,11 @@ import axios from '../__mocks__/axios';
 import { JMAP, MASKED_EMAIL_CALLS } from '../constants';
 import { InvalidArgumentError } from '../error/invalidArgumentError';
 import * as get from '../lib/get';
-import { getByAddress, getById, list } from '../lib/get';
+import { getAllEmails, getEmailByAddress, getEmailById } from '../lib/get';
 import { MaskedEmail } from '../types/maskedEmail';
 
 describe('get', () => {
-  const listSpy = jest.spyOn(get, 'list');
+  const listSpy = jest.spyOn(get, 'getAllEmails');
 
   beforeEach(() => {
     jest.resetAllMocks();
@@ -34,17 +34,19 @@ describe('get', () => {
 
   describe('list', () => {
     it('should reject with InvalidArgumentError if no session is provided', async () => {
-      await expect(list(undefined)).rejects.toThrow(InvalidArgumentError);
+      await expect(getAllEmails(undefined)).rejects.toThrow(
+        InvalidArgumentError
+      );
     });
 
-    it('should list all masked emails', async () => {
+    it('should get all masked emails', async () => {
       axios.post.mockResolvedValue({
         data: {
           methodResponses: [[MASKED_EMAIL_CALLS.get, { list: maskedEmails }]]
         }
       });
 
-      const result = await list(session);
+      const result = await getAllEmails(session);
 
       expect(result).toEqual(maskedEmails);
     });
@@ -52,13 +54,13 @@ describe('get', () => {
 
   describe('getById', () => {
     it('should reject with InvalidArgumentError if no session is provided', async () => {
-      await expect(getById('1', undefined)).rejects.toThrow(
+      await expect(getEmailById('1', undefined)).rejects.toThrow(
         InvalidArgumentError
       );
     });
 
     it('should reject with InvalidArgumentError if no id is provided', async () => {
-      await expect(getById(undefined, session)).rejects.toThrow(
+      await expect(getEmailById(undefined, session)).rejects.toThrow(
         InvalidArgumentError
       );
     });
@@ -72,7 +74,7 @@ describe('get', () => {
         }
       });
 
-      const result = await getById('1', session);
+      const result = await getEmailById('1', session);
 
       expect(result).toEqual(maskedEmails[0]);
     });
@@ -86,7 +88,7 @@ describe('get', () => {
         }
       });
 
-      await expect(getById('masked-email-3', session)).rejects.toThrow(
+      await expect(getEmailById('masked-email-3', session)).rejects.toThrow(
         'No masked email found with id masked-email-3'
       );
     });
@@ -95,7 +97,7 @@ describe('get', () => {
   describe('getByAddress', () => {
     it('should get a masked email by address', async () => {
       listSpy.mockResolvedValue(maskedEmails);
-      const result = await getByAddress('testEmail@test.com', session);
+      const result = await getEmailByAddress('testEmail@test.com', session);
 
       expect(result).toEqual([maskedEmails[0]]);
     });