+++ src/__tests__/set.test.ts
@@ -1,3 +1,5 @@
+import { permanentDeleteFailResponseFixture } from '../__fixtures__/responses/permanentDeleteFailResponse.fixture';
+import { permanentDeleteSuccessResponseFixture } from '../__fixtures__/responses/permanentDeleteSuccessResponse.fixture';
 import axios from '../__mocks__/axios';
 import {
   JMAP,
@@ -5,11 +7,17 @@ import {
   MASKED_EMAIL_CAPABILITY
 } from '../constants';
 import { InvalidArgumentError } from '../error/invalidArgumentError';
-import { disable, enable, remove, update } from '../lib/set';
+import {
+  deleteEmail,
+  disableEmail,
+  enableEmail,
+  permanentlyDeleteEmail,
+  updateEmail
+} from '../lib/set';
 import * as set from '../lib/set';
 import { Options } from '../types/options';
 describe('update', () => {
-  const updateSpy = jest.spyOn(set, 'update');
+  const updateSpy = jest.spyOn(set, 'updateEmail');
 
   beforeEach(() => {
     jest.resetAllMocks();
@@ -23,27 +31,27 @@ describe('update', () => {
     fmAuthToken: 'auth-token-123'
   };
 
-  describe('update', () => {
+  describe('updateEmail', () => {
     it('should reject with InvalidArgumentError if no session is provided', async () => {
       await expect(
-        update('1', undefined, { description: 'test' })
+        updateEmail('1', undefined, { description: 'test' })
       ).rejects.toThrow(InvalidArgumentError);
     });
 
     it('should reject with InvalidArgumentError if no id is provided', async () => {
       await expect(
-        update(undefined, session, { description: 'test' })
+        updateEmail(undefined, session, { description: 'test' })
       ).rejects.toThrow(InvalidArgumentError);
     });
 
     it('should reject with InvalidArgumentError if no options are provided', async () => {
-      await expect(update('1', session, {})).rejects.toThrow(
+      await expect(updateEmail('1', session, {})).rejects.toThrow(
         InvalidArgumentError
       );
     });
     it('should reject with InvalidArgumentError if invalid options are provided', async () => {
       await expect(
-        update('1', session, { invalid: 'invalid' } as Options)
+        updateEmail('1', session, { invalid: 'invalid' } as Options)
       ).rejects.toThrow(InvalidArgumentError);
     });
 
@@ -61,7 +69,7 @@ describe('update', () => {
         }
       });
 
-      const result = await update('1', session, updateOptions);
+      const result = await updateEmail('1', session, updateOptions);
 
       expect(axios.post).toHaveBeenCalledWith(
         'https://api.example.com',
@@ -87,11 +95,11 @@ describe('update', () => {
     });
   });
 
-  describe('remove', () => {
-    it('should remove a masked email', async () => {
+  describe('deleteEmail', () => {
+    it('should delete a masked email', async () => {
       updateSpy.mockResolvedValue({ '1': null });
 
-      const result = await remove('1', session);
+      const result = await deleteEmail('1', session);
 
       expect(updateSpy).toHaveBeenCalledWith('1', session, {
         state: 'deleted'
@@ -100,11 +108,67 @@ describe('update', () => {
     });
   });
 
-  describe('disable', () => {
+  describe('permanentlyDeleteEmail', () => {
+    it('should reject with InvalidArgumentError if no session is provided', async () => {
+      await expect(permanentlyDeleteEmail('1', undefined)).rejects.toThrow(
+        InvalidArgumentError
+      );
+    });
+
+    it('should reject with InvalidArgumentError if no id is provided', async () => {
+      await expect(permanentlyDeleteEmail(undefined, session)).rejects.toThrow(
+        InvalidArgumentError
+      );
+    });
+
+    it('should permanently delete a masked email', async () => {
+      axios.post.mockResolvedValue({
+        data: permanentDeleteSuccessResponseFixture
+      });
+
+      const result = await permanentlyDeleteEmail('masked-81873752', session);
+
+      expect(axios.post).toHaveBeenCalledWith(
+        'https://api.example.com',
+        {
+          using: [JMAP.CORE, MASKED_EMAIL_CAPABILITY],
+          methodCalls: [
+            [
+              MASKED_EMAIL_CALLS.set,
+              { accountId: 'account1', destroy: ['masked-81873752'] },
+              'a'
+            ]
+          ]
+        },
+        {
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: 'Bearer auth-token-123'
+          }
+        }
+      );
+
+      expect(result).toEqual(['masked-81873752']);
+    });
+
+    it('should throw an error if the email could not be deleted', async () => {
+      axios.post.mockResolvedValue({
+        data: permanentDeleteFailResponseFixture
+      });
+
+      await expect(
+        permanentlyDeleteEmail('masked-81873752', session)
+      ).rejects.toThrow(
+        'Only masked emails that have not received email can be destroyed'
+      );
+    });
+  });
+
+  describe('disableEmail', () => {
     it('should disable a masked email', async () => {
       updateSpy.mockResolvedValue({ '1': null });
 
-      const result = await disable('1', session);
+      const result = await disableEmail('1', session);
 
       expect(updateSpy).toHaveBeenCalledWith('1', session, {
         state: 'disabled'
@@ -113,11 +177,11 @@ describe('update', () => {
     });
   });
 
-  describe('enable', () => {
+  describe('enableEmail', () => {
     it('should enable a masked email', async () => {
       updateSpy.mockResolvedValue({ '1': null });
 
-      const result = await enable('1', session);
+      const result = await enableEmail('1', session);
 
       expect(updateSpy).toHaveBeenCalledWith('1', session, {
         state: 'enabled'